# Time_Management_Project/services/back-main/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Time Management Project",
    "description": "Website for tracking the working hours of employees of an IT company",
    "version": "1.0.0"
  },
  "host": "localhost:9000",
...
  "schemes": "http",
  "paths": {
...
    "/changing_dates/get/{id_task}": {
      "get": {
        "tags": ["Changing Dates"],
        "summary": "Returns a changing dates by task.",
        "description": "Endpoint to get a changing dates by task.",
        "parameters": [
          {
            "name": "id_task",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/changing_dates/delete/{changing_date_id}": {
      "get": {
        "tags": ["Changing Dates"],
        "summary": "Delete a changing dates by id.",
        "description": "Endpoint to delete a changing dates by id.",
        "parameters": [
          {
            "name": "changing_date_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/changing_dates/put": {
      "post": {
        "tags": ["Changing Dates"],
        "summary": "Put changing_dates.",
        "description": "Endpoint put changing_dates. cause: string; deadlineBefore: date-time; deadlineAfter: date-time; taskId: integer",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/changing_dates/update/{changing_date_id}": {
      "post": {
        "tags": ["Changing Dates"],
        "summary": "Put changing_dates.",
        "description": "Endpoint put changing_dates. cause: string",
        "parameters": [
          {
            "name": "changing_date_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}